版本号更新为v3.8.1
改动:
速度显示现在能够正确显示MB/s与KB/s
设置窗口→插件管理现在使用了新的界面
取消了启用/禁用插件的功能，所有插件(无论是否已在旧版本中被禁用)都会被启用
设置窗口在通过指定参数调用时会显示精简模式

修复:
修复了即时搜索区分大小写的问题
修复了自动更新后无法启动新版本的问题
修复了优酷解析器无法应用代理服务器设置的问题
修复了等待开始的任务停止时间过长的问题
修复了读取任务信息时字幕文件保存位置错误的问题

+++ AcDown/AcDown/Component/Updater.cs
@@ -118,6 +118,9 @@ namespace Kaedei.AcDown.Component
       {
          string file = filePath.Replace("\"", "");
          //拷贝并覆盖同名文件
+         if (File.Exists(tempFile2))
+            File.Copy(tempFile2, file, true);
+         //拷贝并覆盖同名文件
          if (File.Exists(tempFile))
             File.Copy(tempFile, file, true);
       }
@@ -127,7 +130,14 @@ namespace Kaedei.AcDown.Component
       /// </summary>
       public void DeleteTempFile()
       {
-         File.Delete(tempFile);
+         try
+         {
+            if (File.Exists(tempFile))
+               File.Delete(tempFile);
+            if (File.Exists(tempFile2))
+               File.Delete(tempFile2);
+         }
+         catch { }
       }
 
    }
